Updated error messages and moved a util method from import_job_tasks
